Add popup functionality
In this commit is implemented the required js functions so that the
content in the Featured Project and Projects section is rendered using
Javascript. There is an array which contains the information in all
the projects. The first element of the array will appear in the
Featured project section. The rest will go to the Projects section.
If the array grows the Projects section will continue to add more cards
as needed.

Massive edit on the styling for the popup functionality. Add modal
container to accomplish this end.